Factor code to use a simple wrapper lexer
Handling indentation, logical lines, duplicate line removal, etc. seems
to result in awful code if all of it is done in the lexer, so this
commit begins a refactor to run the lexer on each logical line, with a
separate custom lexer for handling the lines themselves.

Also add a test for duplicate line removal, currently failing.