Add basic initial type objects
We now create a type context that gets passed to the parser and
evaluator to deal with type objects. Since types are no longer
identified by strings, this means that test code now also needs access
to this type context to create expected ASTs and values. Currently we
abuse the 'data' field on objects to identify the types, but soon they
will contain actual methods.